+++ src/chart/mesh.rs
@@ -7,8 +7,8 @@ use crate::coord::{MeshLine, Ranged, RangedCoord};
 use crate::drawing::backend::DrawingBackend;
 use crate::drawing::DrawingAreaErrorKind;
 use crate::style::{
-    AsRelative, Color, FontDesc, FontFamily, IntoTextStyle, RGBColor, ShapeStyle, SizeDesc,
-    TextStyle,
+    AsRelative, Color, FontDesc, FontFamily, FontStyle, IntoTextStyle, RGBColor, ShapeStyle,
+    SizeDesc, TextStyle,
 };
 
 /// The style used to describe the mesh for a secondary coordinate system.
@@ -332,6 +332,7 @@ where
         let default_label_font = FontDesc::new(
             FontFamily::Serif,
             f64::from((12i32).percent().max(12).in_pixels(&self.parent_size)),
+            FontStyle::Normal,
         );
 
         let mesh_style_1 = self